first testing

+++ bookings/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, get_object_or_404, redirect
+from django.http import HttpResponse
 from .models import User, YogaType, YogaClass, Reservation
 from django.contrib import messages
 from .forms import ReservationForm
@@ -31,58 +32,51 @@ def book(request):
                   "15:00 - 16:00", "16:00 - 17:00",
                   "17:00 - 18:00", "20:00 - 21:00"]
     week_days = get_days(request)
-    yoga_classes = weekly_classes(request)
-    yoga_classes_available_now = []
-    for yoga_class in yoga_classes:
-        if yoga_class.day >= date.today():
-            yoga_classes_available_now.append(yoga_class)
-    print("POSSIBILI CLASSI", yoga_classes_available_now)
-
+    yoga_classes = no_obsolete_classes(request)
+    yoga_classes_available_now = yoga_classes_available(request)
+    # allows classes to be accessable only if not in the past
+    # if method is post save form
     if request.method == 'POST':
         form = ReservationForm(request.POST)
         if form.is_valid():
+            form.instance.member_id = request.user.id
             new_reservation = form.save()
-            if new_reservation.yoga_class in yoga_classes_available_now:
-                if request.user.id == new_reservation.member.id:
-                    current_user = request.user
-                    yoga_class_users_reservations = Reservation.objects.filter(
-                        yoga_class_id=new_reservation.yoga_class_id)
-                    yoga_class_user_reservations = \
-                        yoga_class_users_reservations.filter(
-                            member=current_user)
-                    print("RES USER", yoga_class_user_reservations)
-                    print("QUESYSET", yoga_class_user_reservations.count())
-                    if yoga_class_user_reservations.count() > 1:
-                        messages.error(
-                                request, "You are already booked \
-                                    in for this class!")
-                        print("da cancellare")
-                        new_reservation.delete()
-                    else:
-                        reserved_class_id = new_reservation.yoga_class_id
-                        update_approval(request, new_reservation)
-                        print(new_reservation.id)
-                        print("MEMBRO NEWRES", new_reservation.member)
-                        if new_reservation.approved:
-                            reduce_available_spaces(
-                                request, reserved_class_id)
-                            return redirect('reservations')
-                        else:
-                            print("fullybooked")
-                            reservation = get_object_or_404(
-                                Reservation, id=new_reservation.id)
-                            messages.error(
-                                request, 'Unfortunately the class is \
-    fully booked, choose another class!')
-                            new_reservation.delete()
-                else:
-                    messages.warning(
-                        request, "To book a class with a different account \
-login with its details")
-                    new_reservation.delete()
-            else:
-                messages.error(request, "This class is not longer available")
-                new_reservation.delete()
+            valid_reservation(
+                request, new_reservation, yoga_classes_available_now)
+            # if new_reservation.yoga_class in yoga_classes_available_now:
+            #     # if request.user.id == new_reservation.member.id:
+            #     check_double_booking(request, new_reservation)
+                # current_user = request.user
+                # yoga_class_users_reservations = Reservation.objects.filter(
+                #     yoga_class_id=new_reservation.yoga_class_id)
+                # yoga_class_user_reservations = \
+                #     yoga_class_users_reservations.filter(member=current_user)
+                # if yoga_class_user_reservations.count() > 1:
+                #     messages.error(
+                #             request, "You are already booked \
+                #                 in for this class!")
+                #     new_reservation.delete()
+                # else:
+                #     reserved_class_id = new_reservation.yoga_class_id
+
+                # ---updated_reservation = update_approval(
+                #     request, new_reservation)
+
+                # ---fully_booked(request, updated_reservation)
+#                     if new_reservation.approved:
+#                         reduce_available_spaces(
+#                             request, reserved_class_id)
+#                         return redirect('reservations')
+#                     else: 
+#                         reservation = get_object_or_404(
+#                             Reservation, id=new_reservation.id)
+#                         messages.error(
+#                             request, 'Unfortunately the class is \
+# fully booked, choose another class!')
+#                         new_reservation.delete()
+            # else:
+            #     messages.error(request, "This class is not longer available")
+            #     new_reservation.delete()
     form = ReservationForm()
     context = {
         'time_slots': time_slots,
@@ -93,15 +87,63 @@ login with its details")
     return render(request, 'book_class.html', context)
 
 
-def delete_reservation(request, reservation_id):
+def get_days(request):
+    today = date.today()
+    start = today - timedelta(days=today.weekday())
+    end = start + timedelta(days=13)
+    dates = [start + timedelta(days=i) for i in range((end-start).days+1)]
+    return (dates)
+
+
+def no_obsolete_classes(request):
+    week_days = get_days(request)
+    yoga_classes = YogaClass.objects.filter(status=1)
+    if yoga_classes:
+        for yoga_class in yoga_classes:
+            if yoga_class.day < week_days[0]:
+                yoga_class.delete()
+    print("DA INIZIO SETTIMANA", yoga_classes)
+    return yoga_classes
+
+
+def yoga_classes_available(request):
+    yoga_classes = no_obsolete_classes(request)
+    yoga_classes_available = []
+    # allows classes to be accessable only if not in the past
+    for yoga_class in yoga_classes:
+        if yoga_class.day >= date.today():
+            yoga_classes_available.append(yoga_class)
+    return yoga_classes_available
+
+
+def valid_reservation(request, reservation, bookable_classes):
+    if reservation.yoga_class in bookable_classes:
+        check_double_booking(request, reservation)
+    else:
+        messages.error(request, "This class is not longer available")
+        new_reservation.delete()
+
+
+def check_double_booking(request, reservation):
+    current_user = request.user
+    yoga_class_users_reservations = Reservation.objects.filter(
+        yoga_class_id=reservation.yoga_class_id)
+    yoga_class_user_reservations = \
+        yoga_class_users_reservations.filter(member=current_user)
+    if yoga_class_user_reservations.count() > 1:
+        messages.error(
+                request, "You are already booked \
+                    in for this class!")
+        reservation.delete()
+    else:
+        reserved_class_id = reservation.yoga_class_id
+        updated_reservation = update_approval(
+                        request, reservation.id)
+        fully_booked(request, updated_reservation.id)
+
+
+def update_approval(request, reservation_id):
     reservation = get_object_or_404(Reservation, id=reservation_id)
-    reserved_class_id = reservation.yoga_class_id
-    reservation.delete()
-    increase_available_spaces(request, reserved_class_id)
-    return redirect('reservations')
-
-
-def update_approval(request, reservation):
     reserved_class_id = reservation.yoga_class_id
     queryset = YogaClass.objects.filter(status=1)
     chosen_yoga_class = get_object_or_404(queryset, id=reserved_class_id)
@@ -111,6 +153,22 @@ def update_approval(request, reservation):
     else:
         reservation.approved = False
         reservation.save()
+    return redirect('book')
+    # return reservation
+
+
+def fully_booked(request, reservation_id):
+    reservation = get_object_or_404(Reservation, id=reservation_id)
+    if reservation.approved:
+        reduce_available_spaces(
+            request, reservation.yoga_class_id)
+        return redirect('reservations')
+    else:
+        messages.error(
+            request, 'Unfortunately the class is \
+fully booked, choose another class!')
+        reservation.delete()
+        return redirect('book')
 
 
 def reduce_available_spaces(request, chosen_class_id):
@@ -122,6 +180,15 @@ def reduce_available_spaces(request, chosen_class_id):
     chosen_yoga_class.save()
     print("CLASSE ID E SPAZIO-", chosen_yoga_class.id,
           chosen_yoga_class.available_spaces)
+    return redirect('reservations')
+
+
+def delete_reservation(request, reservation_id):
+    reservation = get_object_or_404(Reservation, id=reservation_id)
+    reserved_class_id = reservation.yoga_class_id
+    reservation.delete()
+    increase_available_spaces(request, reserved_class_id)
+    return redirect('reservations')
 
 
 def increase_available_spaces(request, chosen_class_id):
@@ -133,22 +200,4 @@ def increase_available_spaces(request, chosen_class_id):
     chosen_yoga_class.save()
     print("CLASSE ID E SPAZIO+", chosen_yoga_class.id,
           chosen_yoga_class.available_spaces)
-
-
-def get_days(request):
-    today = date.today()
-    start = today - timedelta(days=today.weekday())
-    end = start + timedelta(days=13)
-    dates = [start + timedelta(days=i) for i in range((end-start).days+1)]
-    return (dates)
-
-
-def weekly_classes(request):
-    week_days = get_days(request)
-    yoga_classes = YogaClass.objects.filter(status=1)
-    if yoga_classes:
-        for yoga_class in yoga_classes:
-            if yoga_class.day < week_days[0]:
-                yoga_class.delete()
-    print("DA INIZIO SETTIMANA", yoga_classes)
-    return yoga_classes
+    return redirect('reservations')
